Remove out of context changes
