+++ WooCommerce/Classes/ViewRelated/Dashboard/Analytics Hub/Time Range/Range Data Generation/AnalyticsHubDayRangeData.swift
@@ -1,6 +1,6 @@
 import Foundation
 
-struct AnalyticsHubTodayRangeData: AnalyticsHubTimeRangeData {
+struct AnalyticsHubDayRangeData: AnalyticsHubTimeRangeData {
     let currentDateStart: Date?
     let currentDateEnd: Date?
     let previousDateStart: Date?
