+++ WooCommerce/Classes/ViewRelated/Dashboard/Analytics Hub/Time Range/Range Data Generation/AnalyticsHubYearRangeData.swift
@@ -1,6 +1,6 @@
 import Foundation
 
-struct AnalyticsHubYearToDateRangeData: AnalyticsHubTimeRangeData {
+struct AnalyticsHubYearRangeData: AnalyticsHubTimeRangeData {
     let currentDateStart: Date?
     let currentDateEnd: Date?
     let previousDateStart: Date?
